stt-server-py/pyproject.toml
[project]
name = "speech-to-text-python"
version = "0.1.0"
description = "Python components for speech-to-text application"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24,<2.3", # Changed from >=2.3.1
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "openai-whisper>=20231117",
    "soundfile>=0.12.1",
    "numba>=0.58.0",
    "aiohttp>=3.12.15",
    "aiofiles>=24.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["config", "transcribe", "whisper_wrapper", "llm_processor"]
